refactor:  test시에 자주 사용되는  함수들 testUtils/index.js에 옮김
